Folded linode_api.objects.account submodule into flat file
