--- linode_api/objects/account/__init__.py
@@ -1,9 +1,0 @@
-from .user import User
-from .event import Event
-from .oauth_client import OAuthClient
-from .settings import AccountSettings
-from .user_grant import UserGrants
-from .invoice import Invoice
-from .invoiceitem import InvoiceItem
-from .payment import Payment
-from .account import Account
