--- linode_api/objects/account/payment.py
@@ -1,13 +1,0 @@
-from __future__ import absolute_import
-
-from linode_api.objects import Base, Property
-
-
-class Payment(Base):
-    api_endpoint = "/account/payments/{id}"
-
-    properties = {
-        "id": Property(identifier=True),
-        "date": Property(is_datetime=True),
-        "amount": Property(),
-    }
